Create unit test GH workflow (#134)
This workflow will run the py-marqo unit tests using the `tox` command.
This workflow pulls a Marqo image and runs it. Py-marqo then connects to the 
running container for the tests. 

Unless otherwise specified, the Marqo version that is used for this test will be 
that specified by py-marqo's `marqo.version.__marqo_version__`

Workflow runs every merge to mainline, update to PR request, and can be 
manually run via workflow dispatch.